ai-provider/dataset_v2/requirements.dataset_v2.txt
kaggle
pyyaml
requests
pillow
opencv-python-headless
ultralytics
